+++ src/TRSConversion/Formats/ARI/Unparse/Problem/MetaInfo.hs
@@ -28,7 +28,7 @@ import TRSConversion.Unparse.Utils (filterEmptyDocs)
 -- >>> unparseAriMetaInfo emptyMetaInfo
 -- Nothing
 unparseAriMetaInfo :: MetaInfo -> Doc ann
-unparseAriMetaInfo (MetaInfo cs ds orig sub) =
+unparseAriMetaInfo (MetaInfo {comment = cs, doi= ds, origin = orig, submitted = sub, copsNum = cn}) =
   if null metaBlocks
     then mempty
     else vsep metaBlocks
@@ -37,8 +37,9 @@ unparseAriMetaInfo (MetaInfo cs ds orig sub) =
     metaBlocks =
       filterEmptyDocs $
         maybe [] (map (metaLine "author")) sub
-        ++ [ maybe emptyDoc (metaLine "origin") orig
-           , maybe emptyDoc (metaLine "doi") ds
+        ++ [ maybe emptyDoc (metaLine "doi") ds
+           , maybe emptyDoc (metaLine "origin") orig
+           , maybe emptyDoc (metaLine "cops") cn
            ]
         ++ maybe [] (map commentLine) cs